Add GHC warning flags to bazel build; revert Generic change.
`-Werror` is fine here because we control the exact version of GHC we are
building with (currently 8.2.2), so there is no risk of getting different
warnings on different installations.

Empty data type constructors are now again serialised as numbers instead
of arrays. This matches Binary and hs-msgpack's tests. We can reconsider
if/when the tests are updated in concert with this change.